UI set

+++ resources/views/users/index.blade.php
@@ -1,6 +1,5 @@
 @extends('layouts.app')
 
-
 @section('content')
 
 @if ($message = Session::get('success'))
@@ -50,11 +49,19 @@
               @endif
             </td>
             <td>
-              <a class="btn btn-info btn-sm" href="{{ route('users.show',$user->id) }}">Show</a>
-              <a class="btn btn-warning btn-sm" href="{{ route('users.edit',$user->id) }}">Edit</a>
+              <a class="btn btn-info btn-sm" href="{{ route('users.show',$user->id) }}">
+                Show
+              </a>
+              @can('user-edit')
+              <a class="btn btn-warning btn-sm" href="{{ route('users.edit',$user->id) }}">
+                Edit
+              </a>
+              @endcan
+              @can('user-delete')
                 {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
                 {!! Form::close() !!}
+              @endcan
             </td>
           </tr>
         @endforeach
